IcqEventsListener handlers separated on two new listeners
